+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Adding the `helpers/random` namespace.
 * Adding the `keyers/fingerprint` namespace.
+* Adding the `keyers/ngram-fingerprint` namespace.
 * Adding the `tag/averaged-perceptron` namespace.
 * Adding the `parsers/brown` namespace.
 * Adding the `parsers/conll` namespace.